visp_mimo : Fix mimo mbox crash druing mimo_video probe.
Mbox crash observed during mimo_video probe.
This crash is observed after 32bit variabled are converted to
64-bit in visp/include/media_device.h

Issue is resolved after adding same change in
visp_mimo/include/media_device.h file.

+++ visp_mimo/include/media_device.h
@@ -228,7 +228,7 @@ typedef struct {
 } MediaFmt;
 
 typedef struct {
-    uint32_t DmaAddr;
+    uint64_t DmaAddr;
 	uint32_t DmaSize;
 } MediaPlane;
